+++ src/components/FormPanel/FormPanel.test.tsx
@@ -2135,7 +2135,7 @@ describe('Panel', () => {
        */
       const eventBus = new EventBusSrv();
 
-      it('Should make initial request on dashboard refresh', async () => {
+      it('Should not re-fetch on dashboard refresh when sync is enabled', async () => {
         const data = {
           state: LoadingState.Done,
         };
@@ -2167,7 +2167,7 @@ describe('Panel', () => {
         });
 
         /**
-         * Check if fetch is called again
+         * Verify fetch is not called again
          */
         expect(fetch).toHaveBeenCalledTimes(1);
       });